metaget: mg with extstore and quiet mode
mg quiet mode was failing before, as that'll end up in transmit() with
zero byte iovecs... and transmit() was looking for res > 0.

however, for an error res must be -1 and we simply did a fall through
for the zero case.

I can't think of when we'd want to close the connection when
accidentally writing back zero bytes, but now I'm nervous there's
something doing it deliberately somewhere.