Fix over-freeing in internal object cache
When returning item to the cache and the items allocated exceeded cache->limit, we should free the item and not put it to the cache.

The current code will free any item that tried returned to the cache when cache->total < cache->limit which is true in most cases if the total objects has never exceeded the limit.