watcher.t: reduce flakiness
logger seems to be getting POLLHUP spuriously on slower platforms.

I haven't tracked it down but it's either running poll() tons of times
trying to write to the socket or the way perl is handling replacing a
file descriptor (overwriting $watcher which closes old and creates new
socket).

instead wait for the writes to flush which probably fixes other issues
regardless, and don't overwrite the socket object.