logger: fix spurious watcher hangups
of course the previous commit didn't actually fix the bug: during
poll() call we weren't filtering for the right watcher when we were
polling a specific one. This could lead to reading a POLLHUP on the
wrong socket and other issues.

hopefully that solves some mysteries with the watcher system.